+++ graylog2-server/src/main/java/org/graylog2/inputs/extractors/SplitAndIndexExtractor.java
@@ -35,7 +35,7 @@ public class SplitAndIndexExtractor extends Extractor {
     public SplitAndIndexExtractor(MetricRegistry metricRegistry,
                                   String id,
                                   String title,
-                                  int order,
+                                  long order,
                                   CursorStrategy cursorStrategy,
                                   String sourceField,
                                   String targetField,
